refactor: service, repository 분리
