+++ src/main/java/chess/Application.java
@@ -1,11 +1,16 @@
 package chess;
 
 import chess.controller.ChessController;
-import chess.dao.DbChessGameDao;
+import chess.repository.ChessGameDao;
+import chess.repository.DbChessGameDao;
+import chess.service.ChessService;
 
 public final class Application {
     public static void main(String[] args) {
-        ChessController chessController = new ChessController(new DbChessGameDao());
+        ChessGameDao chessGameDao = new DbChessGameDao();
+        ChessService chessService = new ChessService(chessGameDao);
+        ChessController chessController = new ChessController(chessService);
+
         chessController.run();
     }
 }
