+++ src/main/java/chess/controller/ChessController.java
@@ -1,32 +1,23 @@
 package chess.controller;
 
-import chess.dao.ChessGameDao;
 import chess.domain.chessGame.ChessBoard;
-import chess.domain.chessGame.ChessBoardGenerator;
-import chess.domain.chessGame.Referee;
-import chess.domain.piece.Color;
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
 import chess.dto.request.CommandDto;
 import chess.dto.response.ChessBoardDto;
 import chess.dto.response.StatusDto;
+import chess.service.ChessService;
 import chess.utils.BoardToString;
 import chess.view.GameCommand;
 import chess.view.InputView;
 import chess.view.OutputView;
 
-import java.util.Map;
+import java.util.List;
 
 public final class ChessController {
 
-    private static final String BLACK_TEAM = "검은색";
-    private static final String WHITE_TEAM = "흰색";
-    private static final String DRAW = "무승부";
+    private final ChessService chessService;
 
-    private final ChessGameDao chessGameDao;
-
-    public ChessController(ChessGameDao chessGameDao) {
-        this.chessGameDao = chessGameDao;
+    public ChessController(ChessService chessService) {
+        this.chessService = chessService;
     }
 
     public void run() {
@@ -38,39 +29,18 @@ public final class ChessController {
     }
 
     private void startGame() {
-        ChessBoard chessBoard = setUpChessBoard();
+        ChessBoard chessBoard = chessService.setUpChessBoard();
         showChessBoardStatus(chessBoard);
-        checkKingAlive(chessBoard);
-        try {
-            repeat(() -> playGame(chessBoard));
-        } catch (IllegalStateException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    private ChessBoard setUpChessBoard() {
-        ChessBoard chessBoard = chessGameDao.select();
-        if (chessBoard != null) {
-            return chessBoard;
-        }
-        ChessBoardGenerator generator = new ChessBoardGenerator();
-        chessBoard = new ChessBoard(generator.generate(), Color.WHITE);
-        return chessBoard;
+        repeat(() -> playGame(chessBoard));
     }
 
     private void showChessBoardStatus(ChessBoard chessBoard) {
-        Map<Position, Piece> chessBoardStatus = chessBoard.getChessBoard();
-        String turn = chessBoard.getTurn();
-        ChessBoardDto chessBoardDto = ChessBoardDto.of(BoardToString.convert(chessBoardStatus));
+        List<String> convertedBoard = BoardToString.convert(chessBoard.getChessBoard());
+        ChessBoardDto chessBoardDto = ChessBoardDto.of(convertedBoard);
         OutputView.printChessBoard(chessBoardDto);
-        OutputView.printTurn(turn);
-    }
 
-    private void checkKingAlive(ChessBoard chessBoard) {
-        if (chessBoard.isKingDead()) {
-            chessGameDao.delete(chessBoard);
-            OutputView.printKingDeadMessage();
-        }
+        String turn = chessBoard.getTurn();
+        OutputView.printTurn(turn);
     }
 
     private void repeat(Runnable runnable) {
@@ -83,75 +53,30 @@ public final class ChessController {
     }
 
     private void playGame(ChessBoard chessBoard) {
-        checkKingAlive(chessBoard);
-        CommandDto commandDto = InputView.readPlayingCommand();
-        while (isGameNotEnd(commandDto, chessBoard)) {
+        CommandDto commandDto;
+        do {
+            commandDto = InputView.readPlayingCommand();
             executePlayingCommand(chessBoard, commandDto);
-            commandDto = readPlayingCommandIfKingAlive(chessBoard);
-        }
-        checkKingAlive(chessBoard);
-    }
-
-    private boolean isGameNotEnd(CommandDto commandDto, ChessBoard chessBoard) {
-        return (commandDto.getGameCommand() != GameCommand.END) &&
-                (!chessBoard.isKingDead());
+        } while (isGameNotEnd(commandDto, chessBoard));
     }
 
     private void executePlayingCommand(ChessBoard chessBoard, CommandDto commandDto) {
         GameCommand gameCommand = commandDto.getGameCommand();
         if (gameCommand == GameCommand.MOVE) {
-            executeMoveCommand(chessBoard, commandDto);
+            ChessBoard afterMove = chessService.executeMoveCommand(chessBoard, commandDto);
+            showChessBoardStatus(afterMove);
         }
         if (gameCommand == GameCommand.STATUS) {
-            showStatus(chessBoard);
+            StatusDto statusDto = chessService.getGameStatus(chessBoard);
+            OutputView.printStatus(statusDto);
         }
     }
 
-    private void executeMoveCommand(ChessBoard chessBoard, CommandDto commandDto) {
-        String startInput = commandDto.getStartPosition();
-        String endInput = commandDto.getEndPosition();
-
-        chessBoard.movePiece(Position.of(startInput), Position.of(endInput));
-        showChessBoardStatus(chessBoard);
-        chessGameDao.update(chessBoard);
-    }
-
-    private void showStatus(ChessBoard chessBoard) {
-        Referee referee = new Referee(chessBoard.getChessBoard());
-
-        double blackScore = referee.calculateScore(Color.BLACK);
-        double whiteScore = referee.calculateScore(Color.WHITE);
-        String winner = getWinner(blackScore, whiteScore);
-
-        StatusDto statusDto = StatusDto.of(blackScore, whiteScore, winner);
-        OutputView.printStatus(statusDto);
-    }
-
-    private String getWinner(double blackScore, double whiteScore) {
-        Color winningColor = getWinningColor(blackScore, whiteScore);
-        if (winningColor == Color.BLACK) {
-            return BLACK_TEAM;
+    private boolean isGameNotEnd(CommandDto commandDto, ChessBoard chessBoard) {
+        if (!chessService.checkKingAlive(chessBoard)) {
+            OutputView.printKingDeadMessage();
+            return false;
         }
-        if (winningColor == Color.WHITE) {
-            return WHITE_TEAM;
-        }
-        return DRAW;
-    }
-
-    private Color getWinningColor(double blackScore, double whiteScore) {
-        if (blackScore > whiteScore) {
-            return Color.BLACK;
-        }
-        if (blackScore < whiteScore) {
-            return Color.WHITE;
-        }
-        return null;
-    }
-
-    private CommandDto readPlayingCommandIfKingAlive(ChessBoard chessBoard) {
-        if (chessBoard.isKingDead()) {
-            return CommandDto.of(GameCommand.END);
-        }
-        return InputView.readPlayingCommand();
+        return (commandDto.getGameCommand() != GameCommand.END);
     }
 }
